feat: your groups display + post comments improvements
Implemented group display:
- Added group display on home page and individual post page
- Fixed security for PUT and DELETE comment routes and models
- Debugged report button to show auth popup for signed out users

+++ src/models/PostComments.model.js
@@ -29,9 +29,9 @@ module.exports.insertComments = async function insertComments(data) {
 
 // update Comments by ID (owner only)
 module.exports.updateCommentsByID = async function updateCommentsByID(data) {
-  const VALUES = [data.content, data.id];
+  const VALUES = [data.content, data.id, data.user_id];
   const { rows } = await pool.query(
-    'UPDATE "PostComments" SET "content" = $1 WHERE "id" = $2 RETURNING *',
+    'UPDATE "PostComments" SET "content" = $1 WHERE "id" = $2 AND "user_id" = $3 RETURNING *',
     VALUES
   );
   return rows[0];
@@ -39,7 +39,7 @@ module.exports.updateCommentsByID = async function updateCommentsByID(data) {
 
 // delete a Comment(owner only)
 module.exports.deleteCommentsByID = async function deleteCommentsByID(data) {
-  const VALUES = [data.id];
-  const { rows } = await pool.query('DELETE FROM "PostComments" WHERE "id" = $1 RETURNING *', VALUES);
+  const VALUES = [data.id, data.user_id];
+  const { rows } = await pool.query('DELETE FROM "PostComments" WHERE "id" = $1 AND "user_id" = $2 RETURNING *', VALUES);
   return rows[0];
 };
